Move protocol imports to standalone package.
