+++ src/deps.ts
@@ -3,3 +3,5 @@ export * as path from "https://deno.land/std@0.134.0/path/mod.ts";
 export { ensureDir } from "https://deno.land/std@0.134.0/fs/mod.ts";
 export type { JSONValue } from "https://deno.land/std@0.149.0/encoding/jsonc.ts";
 export { deepMerge } from "https://deno.land/std@0.134.0/collections/deep_merge.ts";
+export { getProtocolInterface } from "https://raw.githubusercontent.com/slackapi/deno-slack-protocols/initial/src/mod.ts";
+export type { Protocol } from "https://raw.githubusercontent.com/slackapi/deno-slack-protocols/initial/src/types.ts";
